+++ client/src/App.tsx
@@ -369,8 +369,7 @@ function App() {
 			disableMutationObserver: false,
 			disable: () =>
 				typeof window !== 'undefined' &&
-				(window.matchMedia('(prefers-reduced-motion: reduce)').matches ||
-					window.innerWidth < 768),
+				window.matchMedia('(prefers-reduced-motion: reduce)').matches,
 		});
 	}, []);